+++ realm/realm-library/src/main/java/io/realm/Realm.java
@@ -17,9 +17,9 @@
 package io.realm;
 
 import android.annotation.TargetApi;
+import android.app.IntentService;
 import android.os.Build;
 import android.os.Looper;
-import android.os.Message;
 import android.util.JsonReader;
 
 import org.json.JSONArray;
@@ -43,6 +43,8 @@ import java.util.Scanner;
 import java.util.Set;
 import java.util.concurrent.Future;
 
+import io.realm.RealmObject;
+import io.realm.RealmQuery;
 import io.realm.annotations.internal.OptionalAPI;
 import io.realm.exceptions.RealmException;
 import io.realm.exceptions.RealmIOException;
@@ -133,11 +135,10 @@ public final class Realm extends BaseRealm {
      * The constructor is private to enforce the use of the static one.
      *
      * @param configuration the {@link RealmConfiguration} used to open the Realm.
-     * @param autoRefresh {@code true} if Realm should auto-refresh. {@code false} otherwise.
      * @throws IllegalArgumentException if trying to open an encrypted Realm with the wrong key.
      */
-    Realm(RealmConfiguration configuration, boolean autoRefresh) {
-        super(configuration, autoRefresh);
+    Realm(RealmConfiguration configuration) {
+        super(configuration);
     }
 
     /**
@@ -236,8 +237,7 @@ public final class Realm extends BaseRealm {
     }
 
     static Realm createAndValidate(RealmConfiguration configuration, ColumnIndices columnIndices) {
-        boolean autoRefresh = Looper.myLooper() != null;
-        Realm realm = new Realm(configuration, autoRefresh);
+        Realm realm = new Realm(configuration);
         long currentVersion = realm.getVersion();
         long requiredVersion = configuration.getSchemaVersion();
         if (currentVersion != UNVERSIONED && currentVersion < requiredVersion && columnIndices == null) {
@@ -790,10 +790,12 @@ public final class Realm extends BaseRealm {
      * you have a large number of object this method is generally faster.
      *
      * @param objects RealmObjects to insert.
+     * @throws IllegalStateException if the corresponding Realm is closed, called from an incorrect thread or not in a
+     * transaction.
      * @see #copyToRealm(Iterable)
      */
     public void insert(Collection<? extends RealmModel> objects) {
-        checkIfValid();
+        checkIfValidAndInTransaction();
         if (objects == null) {
             throw new IllegalArgumentException("Null objects cannot be inserted into Realm.");
         }
@@ -819,10 +821,12 @@ public final class Realm extends BaseRealm {
      * you have a large number of object this method is generally faster.
      *
      * @param object RealmObjects to insert.
+     * @throws IllegalStateException if the corresponding Realm is closed, called from an incorrect thread or not in a
+     * transaction.
      * @see #copyToRealm(RealmModel)
      */
     public void insert(RealmModel object) {
-        checkIfValid();
+        checkIfValidAndInTransaction();
         if (object == null) {
             throw new IllegalArgumentException("Null object cannot be inserted into Realm.");
         }
@@ -846,10 +850,12 @@ public final class Realm extends BaseRealm {
      * you have a large number of object this method is generally faster.
      *
      * @param objects RealmObjects to insert.
+     * @throws IllegalStateException if the corresponding Realm is closed, called from an incorrect thread or not in a
+     * transaction.
      * @see #copyToRealmOrUpdate(Iterable)
      */
     public void insertOrUpdate(Collection<? extends RealmModel> objects) {
-        checkIfValid();
+        checkIfValidAndInTransaction();
         if (objects == null) {
             throw new IllegalArgumentException("Null objects cannot be inserted into Realm.");
         }
@@ -875,10 +881,12 @@ public final class Realm extends BaseRealm {
      * you have a large number of object this method is generally faster.
      *
      * @param object RealmObjects to insert.
+     * @throws IllegalStateException if the corresponding Realm is closed, called from an incorrect thread or not in a
+     * transaction.
      * @see #copyToRealmOrUpdate(RealmModel)
      */
     public void insertOrUpdate(RealmModel object) {
-        checkIfValid();
+        checkIfValidAndInTransaction();
         if (object == null) {
             throw new IllegalArgumentException("Null object cannot be inserted into Realm.");
         }
@@ -1044,7 +1052,7 @@ public final class Realm extends BaseRealm {
      *
      * @param listener the change listener.
      * @throws IllegalArgumentException if the change listener is {@code null}.
-     * @throws IllegalStateException if you try to register a listener from a non-Looper Thread.
+     * @throws IllegalStateException if you try to register a listener from a non-Looper or {@link IntentService} thread.
      * @see io.realm.RealmChangeListener
      * @see #removeChangeListener(RealmChangeListener)
      * @see #removeAllChangeListeners()